+++ pkg/compiler/lib/src/inferrer/powersets/powerset_bits.dart
@@ -589,6 +589,10 @@ class PowersetBitsDomain {
   int get growableListType => _growableListType ??=
       createNonNullExact(commonElements.jsExtendableArrayClass);
 
+  int _mutableArrayType;
+  int get mutableArrayType => _mutableArrayType ??=
+      createNonNullSubtype(commonElements.jsMutableArrayClass);
+
   int get nullType => nullValue;
 
   int get nonNullType => powersetTop & ~nullValue;